+++ src/biodwh2-uniprot/src/main/java/de/unibi/agbi/biodwh2/uniprot/model/NameList.java
@@ -1,0 +1,12 @@
+package de.unibi.agbi.biodwh2.uniprot.model;
+
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
+
+import java.util.List;
+
+public class NameList {
+    @JacksonXmlElementWrapper(useWrapping = false)
+    public List<PersonOrConsortium> person;
+    @JacksonXmlElementWrapper(useWrapping = false)
+    public List<PersonOrConsortium> consortium;
+}