+++ src/biodwh2-uniprot/src/main/java/de/unibi/agbi/biodwh2/uniprot/model/Property.java
@@ -1,0 +1,10 @@
+package de.unibi.agbi.biodwh2.uniprot.model;
+
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
+
+public class Property {
+    @JacksonXmlProperty(isAttribute = true)
+    public String type;
+    @JacksonXmlProperty(isAttribute = true)
+    public String value;
+}