+++ src/biodwh2-uniprot/src/main/java/de/unibi/agbi/biodwh2/uniprot/model/Source.java
@@ -1,0 +1,13 @@
+package de.unibi.agbi.biodwh2.uniprot.model;
+
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
+
+/**
+ * Describes the source of the data using a database cross-reference (or a 'ref' attribute when the source cannot be
+ * found in a public data source, such as PubMed, and is cited only within the UniProtKB entry).
+ */
+public class Source {
+    public DbReference dbReference;
+    @JacksonXmlProperty(isAttribute = true)
+    public Integer ref;
+}